pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gnarum-gsm8k-experiments"
version = "0.3.0"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "optuna>=3.2.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",
    "pynvml>=11.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]